+++ packages/fence_service/lib/src/fence_service_base.dart
@@ -1062,18 +1062,15 @@ class FenceService extends FenceServiceBase {
 
   @override
   Future<StatusResponse> updateOneChain(
-      ServiceCall? call, Chain request) async {
-    if (request.firmId.isEmpty) {
-      throw GrpcError.invalidArgument('request.firmId cannot be empty');
-    }
-    if (request.boutiques.any((b) => b.firmId != request.firmId)) {
+      ServiceCall? call, ChainRequest request) async {
+/*     if (request.boutiques.any((b) => b.firmId != request.firmId)) {
       throw GrpcError.invalidArgument(
           'each boutique.firmId must match the chain.firmId');
     }
     if (request.boutiques.any((b) => b.chainId != request.chainId)) {
       throw GrpcError.invalidArgument(
           'each boutique.chainId must match the chainId');
-    }
+    } */
 
     final userPermission = isMock
         ? userPermissionIfTest ?? UserPermissions()
@@ -1087,13 +1084,25 @@ class FenceService extends FenceServiceBase {
           'user does not have right to update chain');
     }
     if (userPermission.isFirmAndChainAccessible(
-            request.firmId, request.chainId) ==
+            userPermission.firmId, request.chainId) ==
         false) {
       throw GrpcError.permissionDenied(
-          'user cannot access data from firm ${request.firmId} or chain ${request.chainId}');
+          'user cannot access data from firm ${userPermission.firmId} or chain ${request.chainId}');
     }
     _db.isConnected ? null : await _db.open();
-    return await _updateOneChainDBExec(request);
+    await boutiqueCollection.update(
+      where.eq('firmId', userPermission.firmId).eq('chainId', request.chainId),
+      ModifierBuilder()
+          .set('name', request.name)
+          .set('lastUpdatedByuserId', userPermission.userId)
+          .set(
+            'lastUpdateTimestampUTC',
+            DateTime.now().toUtc().timestampProto.toProto3Json(),
+          ),
+    );
+    return StatusResponse()
+      ..type = StatusResponse_Type.UPDATED
+      ..timestamp = DateTime.now().timestampProto;
   }
 
   Future<StatusResponse> _updateOneChainDBExec(Chain chain) async {
@@ -1563,7 +1572,8 @@ class FenceService extends FenceServiceBase {
   }
 
   @override
-  Future<StatusResponse> deleteOneChain(ServiceCall call, Chain request) async {
+  Future<StatusResponse> deleteOneChain(
+      ServiceCall call, ChainRequest request) async {
     final userPermission = isMock
         ? userPermissionIfTest ?? UserPermissions()
         : call.bearer.userPermissions;