+++ packages/protos/protos_weebi/lib/grpc.dart
@@ -1,2 +1,2 @@
 export 'package:grpc/grpc.dart';
-export 'package:protos_weebi/src/service_call_impl.dart';
+export 'package:grpc/grpc_web.dart';